refactor: remove SP1/ZK infrastructure (Phase 1 cleanup)
- Delete circuits/ directory (SP1 word count circuit + build scripts)
- Delete sdk/vkey_hash.txt (SP1 verification key)
- Delete sdk/src/prover.ts (SP1 prover wrapper)
- Delete programs/covenant/src/instructions/submit_completion.rs (ZK verification)
- Delete app/app/proof/, app/app/verify/, app/app/api/proof/, app/app/api/verify/
- Remove sp1-solana dep from programs/covenant/Cargo.toml
- Replace app/lib/sp1-circuit.ts with app/lib/work-metrics.ts
  - New computeWorkMetrics() does same hashing + counting without ZK framing
  - Keeps executeCircuit as deprecated alias so routes keep compiling
- Update all API route imports: @/lib/sp1-circuit -> @/lib/work-metrics
- Remove proof/verify links from NavBar (top nav + More dropdown)
- Dashboard: replace /verify link with /taker
- JobCard: replace 'View Proof' link with 'View Delivery' (/job/[id])
- Stub programs/covenant/src/lib.rs to ping instruction
  (full rewrite lands in Phase 2)
