--- app/lib/sp1-circuit.ts
@@ -1,110 +1,0 @@
-import crypto from "crypto";
-
-export interface SP1ExecutionResult {
-  verified: boolean;
-  category: string;
-  wordCount: number;
-  minWords: number;
-  textHash: string;
-  computedHash: string;
-  hashMatch: boolean;
-  quantityPass: boolean;
-  categoryProof: string;
-  cycleCount: number;
-  executionTime: number;
-  publicValues: {
-    minWords: number;
-    textHash: string;
-    category: string;
-  };
-}
-
-/**
- * Replicates the exact SP1 word-count circuit logic with category-specific verification:
- * 1. Compute SHA-256(text)
- * 2. Assert hash matches expected
- * 3. Count words/items based on category
- * 4. Assert count >= minWords
- * 5. Commit public values
- */
-export function executeCircuit(text: string, minWords: number, category: string = "text_writing"): SP1ExecutionResult {
-  const startTime = performance.now();
-
-  // Step 1: SHA-256 hash
-  const computedHash = crypto.createHash("sha256").update(text, "utf8").digest("hex");
-
-  // Step 2: This IS the expected hash (in real SP1, the expected hash is a public input
-  // that must match. Here we compute it ourselves, same as the circuit does.)
-  const textHash = computedHash;
-  const hashMatch = true;
-
-  // Step 3: Category-specific quantity check
-  let wordCount: number;
-  let quantityPass: boolean;
-  let categoryProof: string;
-
-  switch (category) {
-    case "data_labeling": {
-      // Count JSON items or line-separated items
-      try {
-        const parsed = JSON.parse(text);
-        wordCount = Array.isArray(parsed) ? parsed.length : Object.keys(parsed).length;
-      } catch {
-        wordCount = text.trim().split("\n").filter(l => l.trim()).length;
-      }
-      quantityPass = wordCount >= minWords; // minWords = minItems for this category
-      categoryProof = `Proved ${wordCount} labeled items >= ${minWords} minimum`;
-      break;
-    }
-    case "translation": {
-      wordCount = text.trim().split(/\s+/).filter(w => w.length > 0).length;
-      quantityPass = wordCount >= minWords;
-      categoryProof = `Proved translated output contains ${wordCount} words >= ${minWords} minimum`;
-      break;
-    }
-    case "code_review": {
-      wordCount = text.trim().split(/\s+/).filter(w => w.length > 0).length;
-      quantityPass = wordCount >= minWords;
-      categoryProof = `Proved code review analysis contains ${wordCount} words >= ${minWords} minimum`;
-      break;
-    }
-    case "bug_bounty": {
-      wordCount = text.trim().split(/\s+/).filter(w => w.length > 0).length;
-      quantityPass = wordCount >= minWords;
-      categoryProof = `Proved security report contains ${wordCount} words >= ${minWords} minimum`;
-      break;
-    }
-    case "design": {
-      wordCount = text.trim().split(/\s+/).filter(w => w.length > 0).length;
-      quantityPass = wordCount >= minWords;
-      categoryProof = `Proved design deliverable description contains ${wordCount} words >= ${minWords} minimum`;
-      break;
-    }
-    default: { // text_writing
-      wordCount = text.trim().split(/\s+/).filter(w => w.length > 0).length;
-      quantityPass = wordCount >= minWords;
-      categoryProof = `Proved text output contains ${wordCount} words >= ${minWords} minimum`;
-      break;
-    }
-  }
-
-  // Step 4: Verified = both checks pass
-  const verified = hashMatch && quantityPass;
-
-  const executionTime = performance.now() - startTime;
-
-  return {
-    verified,
-    category,
-    wordCount,
-    minWords,
-    textHash,
-    computedHash,
-    hashMatch,
-    quantityPass,
-    categoryProof,
-    cycleCount: 237583, // Real SP1 measured cycle count
-    executionTime: Math.round(executionTime * 100) / 100,
-    publicValues: { minWords, textHash, category },
-  };
-}
