--- circuits/word_count/script/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "word-count-script"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-sp1-sdk = "6.0.2"
-sha2 = "0.10"
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
-tokio = { version = "1", features = ["full"] }
-hex = "0.4"
-atty = "0.2"
-bincode = "1"
-
-[build-dependencies]
-sp1-build = "6.0.2"
