+++ programs/covenant/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "covenant"
-version = "0.1.0"
-description = "Solana escrow program for trustless job delivery with ZK proof verification"
+version = "0.2.0"
+description = "Open settlement protocol for AI agents — optimistic escrow with bonded dispute resolution"
 edition = "2021"
 
 [lib]
@@ -19,4 +19,3 @@ idl-build = ["anchor-lang/idl-build", "anchor-spl/idl-build"]
 [dependencies]
 anchor-lang = { version = "0.30.1", features = ["init-if-needed"] }
 anchor-spl = "0.30.1"
-sp1-solana = "0.1.0"
