+++ programs/covenant/src/lib.rs
@@ -1,13 +1,7 @@
+// NOTE: This file will be fully rewritten in Phase 2.
+// Temporarily stubbed during Phase 1 cleanup to keep the build graph consistent
+// after removing submit_completion / ZK.
 use anchor_lang::prelude::*;
-
-pub mod errors;
-pub mod instructions;
-pub mod state;
-
-pub use instructions::create_job::*;
-pub use instructions::accept_job::*;
-pub use instructions::submit_completion::*;
-pub use instructions::cancel_job::*;
 
 declare_id!("HAptQVTwT4AYRzPkvT9UFxGEZEjqVs6ALF295WXXPTNo");
 
@@ -15,24 +9,12 @@ declare_id!("HAptQVTwT4AYRzPkvT9UFxGEZEjqVs6ALF295WXXPTNo");
 pub mod covenant {
     use super::*;
 
-    pub fn create_job(ctx: Context<CreateJob>, amount: u64, spec_hash: [u8; 32], deadline: i64) -> Result<()> {
-        instructions::create_job::handler(ctx, amount, spec_hash, deadline)
+    pub fn ping(_ctx: Context<Ping>) -> Result<()> {
+        Ok(())
     }
+}
 
-    pub fn accept_job(ctx: Context<AcceptJob>, spec_hash: [u8; 32]) -> Result<()> {
-        instructions::accept_job::handler(ctx, spec_hash)
-    }
-
-    pub fn submit_completion(
-        ctx: Context<SubmitCompletion>,
-        proof: Vec<u8>,
-        min_words: u32,
-        text_hash: [u8; 32],
-    ) -> Result<()> {
-        instructions::submit_completion::handler(ctx, proof, min_words, text_hash)
-    }
-
-    pub fn cancel_job(ctx: Context<CancelJob>) -> Result<()> {
-        instructions::cancel_job::handler(ctx)
-    }
+#[derive(Accounts)]
+pub struct Ping<'info> {
+    pub signer: Signer<'info>,
 }
