--- sdk/vkey_hash.txt
@@ -1,1 +1,0 @@
-0x002b54c2ee0f83205f876710bd9bc4cabf71fb0a73d872fb8769dea99e133b9f
